further improvements
- remove unnecessary db clients from E2ET because script consumes data from table and print it to console
- fixed version.property path in commons
- unified way of printing supported script engines
- added missing compositekey.ben.xml demo case
- preparations maven central upload
- delete .classpath because it is ide related
- preparation demo cases for GraalVM Scripts
- removed old logging config

+++ src/test/java/com/rapiddweller/script/IsThreeExpression.java
@@ -27,7 +27,7 @@ class IsThreeExpression implements Expression<Boolean> {
 	@Override
 	public Boolean evaluate(Context context) {
 		Integer candidateValue = (Integer) context.get("_candidate");
-		return (candidateValue != null && candidateValue.intValue() == 3);
+		return (candidateValue != null && candidateValue == 3);
 	}
 
 	@Override
